サービスクラスの実装
