+++ src/application/gameService.ts
@@ -1,0 +1,36 @@
+import { GameGateway } from "../dataaccess/gameGateway";
+import { SquareGateway } from "../dataaccess/squareGateway";
+import { TurnGateway } from "../dataaccess/turnGateway";
+import { connectMySQL } from "../dataaccess/connection";
+import { DARK, INITIAL_BOARD } from "../application/constants";
+
+const gameGateway = new GameGateway();
+const turnGateway = new TurnGateway();
+const squareGateway = new SquareGateway();
+
+export class GameService {
+  async startNewGame() {
+    const now = new Date();
+    const conn = await connectMySQL();
+    try {
+      await conn.beginTransaction();
+      const gameRecord = await gameGateway.insert(conn, now);
+      const turnRecord = await turnGateway.insert(
+        conn,
+        gameRecord.id,
+        0,
+        DARK,
+        now
+      );
+      const squareCount = INITIAL_BOARD.map((line) => line.length).reduce(
+        (v1, v2) => v1 + v2,
+        0
+      );
+      await squareGateway.insertAll(conn, turnRecord.id, INITIAL_BOARD);
+
+      await conn.commit();
+    } finally {
+      await conn.end();
+    }
+  }
+}
