+++ src/presentation/gameRouter.ts
@@ -1,39 +1,11 @@
 import express from "express";
-import { GameGateway } from "../dataaccess/gameGateway";
-import { SquareGateway } from "../dataaccess/squareGateway";
-import { TurnGateway } from "../dataaccess/turnGateway";
-import { connectMySQL } from "../dataaccess/connection";
-import { DARK, INITIAL_BOARD } from "../application/constants";
+import { GameService } from "../application/gameService";
 
 export const gameRouter = express.Router();
 
-const gameGateway = new GameGateway();
-const turnGateway = new TurnGateway();
-const squareGateway = new SquareGateway();
+const gameService = new GameService();
 
 gameRouter.post("/api/games", async (req, res) => {
-  const now = new Date();
-  const conn = await connectMySQL();
-  try {
-    await conn.beginTransaction();
-    const gameRecord = await gameGateway.insert(conn, now);
-    const turnRecord = await turnGateway.insert(
-      conn,
-      gameRecord.id,
-      0,
-      DARK,
-      now
-    );
-    const squareCount = INITIAL_BOARD.map((line) => line.length).reduce(
-      (v1, v2) => v1 + v2,
-      0
-    );
-    await squareGateway.insertAll(conn, turnRecord.id, INITIAL_BOARD);
-
-    await conn.commit();
-  } finally {
-    await conn.end();
-  }
-
+  await gameService.startNewGame();
   res.status(201).end();
 });
